some refatoring
