+++ models/ParserConfiguration.php
@@ -31,21 +31,6 @@ use yii\base\NotSupportedException;
  */
 class ParserConfiguration extends \netis\utils\crud\ActiveRecord
 {
-    public $length;
-    public $delimiter;
-    public $enclosure;
-    public $escape;
-    public $firstCol;
-    public $firstRow;
-    public $sheet;
-    public $header;
-
-    public $csvParserFields = ['length', 'delimiter', 'enclosure', 'escape', 'header'];
-    public $xlsParserFields = ['firstCol', 'firstRow', 'sheet', 'header'];
-
-    const SCENARIO_CSV_PARSER = 'csvParser';
-    const SCENARIO_XLS_PARSER = 'xlsParser';
-
     /**
      * @inheritdoc
      */
@@ -74,12 +59,6 @@ class ParserConfiguration extends \netis\utils\crud\ActiveRecord
                 }
             }],
             [['name', 'model_class', 'parser_class'], 'string', 'max' => 255],
-            [['length'], 'integer', 'on' => self::SCENARIO_CSV_PARSER],
-            [['delimiter', 'enclosure', 'escape'], 'string', 'max' => 1, 'on' => self::SCENARIO_CSV_PARSER],
-            [['header'], 'boolean', 'on' => self::SCENARIO_CSV_PARSER],
-            [['firstCol'], 'string', 'on' => self::SCENARIO_XLS_PARSER],
-            [['firstRow', 'sheet'], 'integer', 'on' => self::SCENARIO_XLS_PARSER],
-            [['header'], 'boolean', 'on' => self::SCENARIO_XLS_PARSER],
         ];
     }
 
@@ -99,14 +78,6 @@ class ParserConfiguration extends \netis\utils\crud\ActiveRecord
             'editor_id' => Yii::t('nineinchnick/sync/models', 'Editor ID'),
             'updated_on' => Yii::t('nineinchnick/sync/models', 'Updated On'),
             'created_on' => Yii::t('nineinchnick/sync/models', 'Created On'),
-            'length' => Yii::t('nineinchnick/sync/models', 'Length'),
-            'delimiter' => Yii::t('nineinchnick/sync/models', 'Delimiter'),
-            'enclosure' => Yii::t('nineinchnick/sync/models', 'Enclosure'),
-            'escape' => Yii::t('nineinchnick/sync/models', 'Escape'),
-            'fistCol' => Yii::t('nineinchnick/sync/models', 'First Column'),
-            'firstRow' => Yii::t('nineinchnick/sync/models', 'First Row'),
-            'sheet' => Yii::t('nineinchnick/sync/models', 'Sheet'),
-            'header' => Yii::t('nineinchnick/sync/models', 'Header'),
         ];
     }
 
@@ -153,15 +124,6 @@ class ParserConfiguration extends \netis\utils\crud\ActiveRecord
             'editor',
             'transactions',
         ];
-    }
-
-    public function scenarios()
-    {
-        return array_merge(parent::scenarios(), [
-            self::SCENARIO_CSV_PARSER => ['name', 'model_class', 'parser_class', 'parser_options', 'length', 'delimiter', 'enclosure', 'escape', 'header'],
-            self::SCENARIO_XLS_PARSER => ['name', 'model_class', 'parser_class', 'parser_options', 'firstCol', 'firstRow', 'sheet', 'header'],
-        ]);
-
     }
 
     /**
